+++ src/main/javascript/Core/AppEvents.js
@@ -20,6 +20,8 @@ export const EVENT_ICON_CHANGED = 'app.icon_changed';
 export const EVENT_BADGECOUNT_CHANGED = 'app.badgecount_changed';
 
 export const EVENT_RESET_SIZE = 'app.reset_size';
+export const EVENT_SHOW_NOTIFICATION = 'app.show_notification';
+
 
 export const events = {
   EVENT_MOUNT,
@@ -38,7 +40,8 @@ export const events = {
   EVENT_REFRESH,
   EVENT_UNLOAD,
 
-  EVENT_RESET_SIZE
+  EVENT_RESET_SIZE,
+  EVENT_SHOW_NOTIFICATION
 };
 
 export const eventNames = Object.keys(events).map(key => events[key]);