+++ src/main/javascript/Core/MessageBroker.js
@@ -5,7 +5,8 @@ import { WidgetMessage } from './WidgetMessage';
 
 const handleRequest = (requestHandler, initialRequest) => new Promise((resolveStart, rejectStart) => requestHandler(resolveStart, rejectStart, initialRequest));
 const createResponseListener = (responseHandler, resolve, reject) => response => responseHandler(resolve, reject, new WidgetMessage(response));
-const createRequestMessage = (eventName, request) => MessageFactory.createRequest(windowProxy.xchild, request);
+const createRequestResponseMessage = (eventName, request) => MessageFactory.createRequestResponse(windowProxy.xchild, request);
+const createFireAndForgetMessage = (eventName, request) => MessageFactory.createFireAndForget(windowProxy.xchild, request);
 
 const sendRequest = (eventName, message) => {
   const { props } = windowProxy.xchild;
@@ -14,16 +15,27 @@ const sendRequest = (eventName, message) => {
 
 const registerResponseListener = (eventName, message, responseListener) => ResponseEventDispatcher.once(eventName, responseListener);
 
-export const createEventListener = (eventName, requestHandler, responseHandler) => {
+export const createRequestResponseEventListener = (eventName, requestHandler, responseHandler) => {
   return function (resolve, reject, initialRequest) {
       const responseListener = createResponseListener(responseHandler, resolve, reject);
 
       handleRequest(requestHandler, initialRequest)
-        .then(request => createRequestMessage(eventName, request))
+        .then(request => createRequestResponseMessage(eventName, request))
         .then(message => { registerResponseListener(eventName, message, responseListener); return message; }  )
         .then(message => sendRequest(eventName, message))
         .catch(e => reject(e))
       ;
+  }
+};
+
+export const createFireAndForgetEventListener = (eventName, requestHandler) => {
+  return function (resolve, reject, initialRequest) {
+
+    handleRequest(requestHandler, initialRequest)
+      .then(request => createFireAndForgetMessage(eventName, request))
+      .then(message => sendRequest(eventName, message))
+      .catch(e => reject(e))
+    ;
   }
 };